# Profile Signature-Aware Client Decoding Techniques