# Old way...
```
Label                                         Count    Total(s)     Avg(s)               Min(s)               Max(s)              
--------------------------------------------------------
DelegateChanged(address,address,address)      1449415  2.312653     0.000001595577       0.000001291744       0.001809665933      
DelegateVotesChanged(address,uint256,uint256) 10516996 17.496441    0.000001663635       0.000001291744       0.001319709234      
Transfer(address,address,uint256)             74049673 119.161042   0.000001609204       0.000001291744       0.008532041684      
--------------------------------------------------------
```

# New way...
```
Label                                         Count    Total(s)     Avg(s)               Min(s)               Max(s)              
--------------------------------------------------------
DelegateChanged(address,address,address)      1449414  2.288299     0.000001578775       0.000001332723       0.001340541989      
DelegateVotesChanged(address,uint256,uint256) 10516995 17.418912    0.000001656263       0.000001290813       0.001096208114      
Transfer(address,address,uint256)             74049672 118.352584   0.000001598286       0.000001290813       0.008739416953      
--------------------------------------------------------
```

...slightly faster! 

Note that we have one less record counted in the "new way", because we have to introspect the first object, and the Profiler class needs a label before it starts.